reducer tests split
