+++ __tests__/reducers/institutions.js
@@ -1,0 +1,30 @@
+import * as types from '../../src/js/constants'
+import excludeTypes from './excludeTypes.js'
+import institutions from '../../src/js/reducers/institutions.js'
+
+describe('institutions reducer', () => {
+  it('should return the initial state on empty action', () => {
+    expect(
+      institutions(undefined, {})
+    ).toEqual({})
+  })
+
+  it('handles REQUEST_INSTITUTIONS', () => {
+    expect(
+      institutions({}, {type: types.REQUEST_INSTITUTIONS})
+    ).toEqual({isFetching: true})
+  })
+
+  it('handles RECEIVE_INSTITUTIONS', () => {
+    expect(
+      institutions({}, {type: types.RECEIVE_INSTITUTIONS, institutions:{a:1}})
+    ).toEqual({isFetching: false, institutions: {a: 1}})
+  })
+
+  it('shouldn\'t modify state on an unknown action type', () => {
+    excludeTypes(types.RECEIVE_INSTITUTIONS, types.REQUEST_INSTITUTIONS)
+      .forEach(v => expect(institutions({}, v))
+        .toEqual({})
+      )
+  })
+})
